Use microG for WebAuthn support in Gecko
Also drop Ads Identifier library dummy to avoid a conflict between
microG library and this dummy.

WebAuthn will work only if microG Service Core is installed on the
device! It will not work with Google Services because Google requires
a cumbersome registration of an app before they allow it to use
WebAuthn API in GMS.

See https://microg.org for more info on microG.