Add NOT supported, NOT recommended secret settings to toggle EME and Widevine (Off by default)
Currently, to play DRM-controlled content, users are enabling EME and Widevine from the `about:config` by setting `media.eme.enabled` and `media.mediadrm-widevinecdm.visible` to `true`. This *technically* works, but the problem is that we remove the EME permission UI in `fenix-liberate.patch` - so if a user enables EME with the `media.eme.enabled` pref (which we know users are), it allows *all* websites to use DRM, without prompting.

So for folks who insist on enabling/using DRM (which I will emphasize is NOT supported or recommended), this adds a way for them to enable it at their discretion, while still being able to control which sites can and can't use it, like vanilla Firefox allows.

This adds two hidden/secret settings that function as follows:

- `Enable Encrypted Media Extensions (EME)` - When enabled, this sets `media.eme.enabled` to true, and it exposes the UI for controlling the DRM site permission.
- `Enable Widevine CDM` - This depends on the `Enable Encrypted Media Extensions (EME)` setting. When enabled, it sets `media.mediadrm-widevinecdm.visible` to `true`.

Signed-off-by: celenity <celenity@celenity.dev>