Set `media.geckoview.autoplay.request.testing` to `0` to re-enable media autoplay in GeckoView
When `media.geckoview.autoplay.request` is set to `true` (which it is for us), the `media.geckoview.autoplay.request.testing` pref controls autoplay, and it takes precedence over the site permission.

Since Phoenix can't control UI settings, I'm setting this to `2` there for next release (which is equivalent to blocking it). This doesn't make sense for us though, in favor of the site permission (which we set to block media autoplay by default with the `default-site-permissions` patch)

Signed-off-by: celenity <celenity@celenity.dev>