fix: use absolute path for APK_IN when signing APK
Builds can generate multiple APK files now. When building APK for a specific CPU ABI a universal APK is also generated. As a result, the output of `ls` command includes paths for two APKs which is then stored in APK_IN. APK_IN is then provided to apksigner, which fails with ENOENT because the input file path is invalid. This commit fixes that issue.

Signed-off-by: Akash Yadav <itsaky01@gmail.com>