feat: Create RS Blocker for Application Services
So, the vast majority of Remote Settings collections and related functionality handled by Gecko directly (hence, they're covered by the `gecko-rs-blocker` patch), but, this is NOT the case for everything. In addition to Gecko's Remote Settings instance, Fenix also has its own (created/managed by Application Services). Currently, use of this AS Remote Settings instance is small - it's use is mainly for Firefox Suggest, search configurations, search telemetry (which we remove the dump for), etc. But, its use may grow in the future, and in general, it makes sense for us to have control over this, the same way we do for Gecko's RS instance. This is also especially important for search configs, since we'll want to provide those locally, but we don't want to fetch them remotely.

For the sake of clarity: Collections blocked/allowed in AS Remote Settings have NO impact on Gecko Remote Settings, and vice versa.

Signed-off-by: celenity <celenity@celenity.dev>