feat: Create search configuration for Mozilla's newer Remote Settings implementation
We're currently using the local implementation for customizing search engines, but it's likely that this implementation will eventually be removed by Mozilla, in favor of the new RS implementation

So once the legacy/existing local implementation is removed, this will ensure we're ready to switch to the new one. Now that we have RS Blocker for Application Services, we can also ensure that only our default/packaged search config is used, instead of some remote config from Mozilla.

Signed-off-by: celenity <celenity@celenity.dev>