feat: Set Gecko preferences for UI settings from Fenix directly, via Mozilla's implementation of the new GeckoPreferenceController API: https://mozilla.github.io/geckoview/javadoc/mozilla-central/org/mozilla/geckoview/GeckoPreferenceController.html
So, this is pretty amazing for us. For one, it will now be significantly easier for us to add UI settings for Gecko prefs. If we want to add a new setting, we no longer have to create GeckoRuntimeSettings and GeckoEngine settings with patches for GeckoView and Android Components - we can simply set them directly (with our new GeckoSettingsBridge). This will make maintenance and updates significantly easier, and allows us to potentially add new UI settings for prefs that we may not have been motivated to before.

This also gets rid of our hidden, unsupported UI settings to control EME and Widevine - because, we now expose the EME UI based on the value of `media.eme.enabled` directly.

In general, this is an overwhelmingly positive change - thank you Mozilla for finally making this possible <3

Signed-off-by: celenity <celenity@celenity.dev>